// backend/src/dataTypes/enums.ts
export enum EnumStatus {
  None = "None",
  SeatIn = "Seat In",
  InWaitingList = "In Waiting List",
  ServiceCompleted = "Service Completed",
}

export enum EnumNotificationUser {
  Self = "Self",
  CanCheckInNow = "Can Check In Now",
  StillInWaiting = "Still In Waiting",
}

export enum EnumCount {
  BookedSeats = "Booked Seats",
  CanCheckInSeats = "Can Check In Seats",
  UsersInWaiting = "Users In Waiting",
}